cli: address review — keep the leaves in place, lazy-import instead of relocating
Per review, undo the two file relocations and reach the modules the way
the rule intends — lazily at the CLI use site — rather than moving them
into a leaf zone:

- `providers/weak-open-model`: moved back; `oauth/status.ts` now
  `await import()`s it inside `noConnectionsMessage` (made async; its one
  caller and unit test updated to await).
- `browser/operation-meta`: moved back. It is consumed synchronously
  while `buildCliProgramTree()` assembles the `browser` subcommand tree,
  so it cannot be a lazy `import()`. It is an execution-free data leaf
  (no Playwright graph), so the single hoisted import carries a scoped
  `eslint-disable-next-line cli/no-daemon-internals` with a comment
  explaining why.

Rule, AGENTS.md, and the other lazifications are unchanged.
