+++ assistant/src/cli/commands/browser.ts
@@ -11,12 +11,18 @@ import { writeFileSync } from "node:fs";
 
 import { type Command, Option } from "commander";
 
+// operation-meta is an execution-free data leaf (no Playwright imports),
+// consumed synchronously while building the `browser` subcommand tree — the
+// program is assembled by the sync `buildCliProgramTree()`, so this can't be a
+// lazy `import()`. Kept out of the heavier `browser/operations` barrel on
+// purpose.
+// eslint-disable-next-line cli/no-daemon-internals
+import { BROWSER_OPERATION_META } from "../../browser/operation-meta.js";
 import type {
   BrowserOperationMeta,
   OperationField,
 } from "../../browser/types.js";
 import { cliIpcCall } from "../../ipc/cli-client.js";
-import { BROWSER_OPERATION_META } from "../../util/browser-operation-meta.js";
 import { registerCommand } from "../lib/register-command.js";
 import { log } from "../logger.js";
