+++ assistant/src/cli/commands/oauth/status.test.ts
@@ -10,25 +10,25 @@ afterEach(() => {
 });
 
 describe("noConnectionsMessage", () => {
-  test("capable models get the terse default", () => {
+  test("capable models get the terse default", async () => {
     process.env.__RESOLVED_MODEL = "claude-opus-4-8";
-    const msg = noConnectionsMessage("google");
+    const msg = await noConnectionsMessage("google");
     expect(msg).toBe(
       "No active connections for google.\n" +
         "Connect with `assistant oauth connect google`.\n",
     );
   });
 
-  test("no resolved model falls back to the terse default", () => {
+  test("no resolved model falls back to the terse default", async () => {
     delete process.env.__RESOLVED_MODEL;
-    expect(noConnectionsMessage("google")).toContain(
+    expect(await noConnectionsMessage("google")).toContain(
       "Connect with `assistant oauth connect google`.",
     );
   });
 
-  test("weak open models are steered to the oauth_connect surface", () => {
+  test("weak open models are steered to the oauth_connect surface", async () => {
     process.env.__RESOLVED_MODEL = "accounts/fireworks/models/minimax-m3";
-    const msg = noConnectionsMessage("google");
+    const msg = await noConnectionsMessage("google");
     expect(msg).toContain('surface_type "oauth_connect"');
     expect(msg).toContain('data.providerKey "google"');
     expect(msg).not.toContain("assistant oauth connect google");
