+++ assistant/src/providers/weak-open-model.ts
